// src/modules/system/cloud-storage/services/file/file-messages/file-messages.enum.ts
export enum FileMessage {
  FILES_NOT_DEFINED = 'Files not defined',
  INVALID_FILE = 'Invalid file',
  FILE_NOT_DEFINED = 'File not defined',
  FILE_NOT_FOUND = 'File not found',
  NAME_NOT_DEFINED = 'File name not defined',
  INVALID_FILE_NAME = 'Invalid file name',
  INVALID_FILE_NAMES = 'Invalid file names',
  DIRECTORY_NOT_DEFINED = 'Directory not defined',
  INVALID_DIRECTORY_NAME = 'Invalid directory name',
}